requirements.txt
peewee
bs4
python-dotenv
tweepy
colorlog
requests
google-cloud-firestore
firebase-admin